Fixes and improvements (#2)
* fixing bugs and adding features

Bug fixed:

- the path by which to collect modules to look for pointers over was
just set to the same as the pytest startdir. This resulted in erroneous
situations where it was picking up modules installed in virtualenv
folders etc. The --pointers-collect option now takes a string as an
argument that defaults to 'src' to indicate the path relative to the
pytest startdir that you actually want to collect the functions from.

Features added:

- add '--pointers-func-min-pass=INT' which can configure what you
consider a pass for each function collected. This maintains the same
default as before where 2 pointer marks are needed to pass and to get
green. Anything that is > 0 but not yet passing is printed as blue.

- add '--pointers-percent-pass=FLOAT' similar to a coverage failure
percentage this exits the testing with exit 1 if the specified
percentage of individual units/functions is not met. This required
moving the report logic into a different hook wrapper.

* adds README documentation, reworks some basics

- add README instructions which are more detailed than before and
explains the new features.
- improved the output formatting and fixes report logic
- refactor the option naming to be consistent with pytest-cov